pyproject.toml
[tool.poetry]
name = "ytsp"
version = "0.1.0"
description = "This is a spotify playlist creator from YouTube videos that transcribe the video using open AI and creates a playlist in Spotify based on the recommendations from the video."
authors = ["José Enrique Pons"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.334"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.107.0"
spotipy = "^2.23.0"
openai = "^1.2.3"
fastapi = "^0.104.1"
uvicorn = "0.24.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
